requirements/base.txt
Django==2.1.7
django-cors-headers==2.4.0
django-model-utils==3.1.2
djangorestframework==3.9.1
Pillow==5.4.1
pytz==2018.9
